+++ server/src/main/java/org/opensearch/common/settings/FeatureFlagSettings.java
@@ -37,6 +37,6 @@ public class FeatureFlagSettings extends AbstractScopedSettings {
         FeatureFlags.WRITEABLE_REMOTE_INDEX_SETTING,
         FeatureFlags.DOC_ID_FUZZY_SET_SETTING,
         FeatureFlags.REMOTE_STORE_MIGRATION_EXPERIMENTAL_SETTING,
-        FeatureFlags.TIERED_CACHING_SETTING
+        FeatureFlags.PLUGGABLE_CACHE_SETTING
     );
 }